added node_container base class
