+++ scripts/scenes/overworld/elements/rings.py
@@ -6,19 +6,19 @@ from typing import TYPE_CHECKING
 
 import pygame
 
-from scripts.core.constants import NodeType
+from scripts.core.base_classes.node_container import NodeContainer
 from scripts.scenes.overworld.elements.node2 import Node2
 
 if TYPE_CHECKING:
     from scripts.core.game import Game
     from typing import List, Tuple, Dict
 
-__all__ = ["Ring"]
+__all__ = ["Rings"]
 
 
-class Ring:
-    def __init__(self, game: Game,  centre: Tuple[int, int], outer_radius: float, num_rings):
-        self.game: Game = game
+class Rings(NodeContainer):
+    def __init__(self, game: Game, centre: Tuple[int, int], outer_radius: float, num_rings: int):
+        super().__init__(game)
         self.outer_radius: float = outer_radius
         self.centre: Tuple[int, int] = centre
         self.num_rings: int = num_rings
@@ -66,7 +66,7 @@ class Ring:
                 y = self.centre[1] + vec.y
 
                 # generate node type
-                node_type = self.get_random_node_type()
+                node_type = self._get_random_node_type()
 
                 # get node icon
                 node_icon = self._get_node_icon(node_type)
@@ -106,51 +106,3 @@ class Ring:
 
                 except KeyError:
                     pass
-
-
-    def _get_node_icon(self, node_type: NodeType) -> pygame.Surface:
-        """
-        Get the node icon from the node type
-        """
-        if node_type == NodeType.COMBAT:
-            node_icon = self.game.assets.get_image("nodes", "combat")
-        elif node_type == NodeType.EVENT:
-            node_icon = self.game.assets.get_image("nodes", "event")
-        elif node_type == NodeType.INN:
-            node_icon = self.game.assets.get_image("nodes", "inn")
-        elif node_type == NodeType.TRAINING:
-            node_icon = self.game.assets.get_image("nodes", "training")
-        else:
-            # node_type == NodeType.UNKNOWN
-            node_icon = self.game.assets.get_image("nodes", "unknown")
-
-        return node_icon
-
-    def get_random_node_type(self, allow_unknown: bool = True) -> NodeType:
-        """
-        Return a random node type
-        """
-        node_weights_dict = self.game.data.config["overworld"]["node_weights"]
-        node_types = [NodeType.COMBAT, NodeType.EVENT, NodeType.INN, NodeType.TRAINING]
-
-        if allow_unknown:
-            node_types.append(NodeType.UNKNOWN)
-
-        node_weights = []
-        try:
-            for enum_ in node_types:
-                name = enum_.name.lower()
-                node_weights.append(node_weights_dict[name])
-
-        except KeyError as key_error:
-            logging.warning(f"generate_map: Node key not found in config file. Defaults used. err:{key_error}")
-
-            # overwrite with default
-            node_weights = []
-            for enum_ in node_types:
-                node_weights.append(0.1)
-
-        node_type = self.game.rng.choices(node_types, node_weights, k=1)[0]
-
-        return node_type
-
