+++ scripts/scenes/overworld/ui.py
@@ -37,7 +37,7 @@ class OverworldUI(UI):
         super().update(delta_time)
 
         if self.game.overworld.state == OverworldState.READY:
-            row_nodes = self.game.overworld.nodes[self.game.overworld.current_node_row]
+            #row_nodes = self.game.overworld.nodes[self.game.overworld.current_node_row]
 
             if self.game.input.states["left"]:
                 self.game.input.states["left"] = False
@@ -90,33 +90,33 @@ class OverworldUI(UI):
             self.game.assets.fonts["warning"].render("Loading...", surface, (10, window_height - 20))
 
         elif overworld_map.state == OverworldState.READY:
-            # get node icon details
-            node_width = overworld_map.nodes[0][0].icon.get_width()
-            node_height = overworld_map.nodes[0][0].icon.get_height()
-            line_offset = 2
-
-            # draw node connections
-            for row in overworld_map.nodes:
-                for node in row:
-                    for connected_node in node.connected_previous_row_nodes:
-                        # get positions
-                        node_centre_x = node.pos[0] + (node_width / 2)
-                        node_top = node.pos[1] - line_offset
-                        connected_node_centre = connected_node.pos[0] + (node_width / 2)
-                        connected_node_bottom = connected_node.pos[1] + node_height + line_offset
-                        # FIXME - are these nodes the wrong way round?
-
-                        pygame.draw.line(
-                            surface,
-                            (255, 255, 255),
-                            (node_centre_x, node_top),
-                            (connected_node_centre, connected_node_bottom),
-                        )
-
-            # draw selection
-            selected_node = overworld_map.nodes[overworld_map.current_node_row][self.selected_node]
-            selected_node_centre_x = selected_node.pos[0] + (node_width / 2)
-            selected_node_centre_y = selected_node.pos[1] + (node_height / 2)
-            pygame.draw.circle(surface, (255, 0, 0), (selected_node_centre_x, selected_node_centre_y), node_width, 2)
+            # # get node icon details
+            # node_width = overworld_map.nodes[0][0].icon.get_width()
+            # node_height = overworld_map.nodes[0][0].icon.get_height()
+            # line_offset = 2
+            #
+            # # draw node connections
+            # for row in overworld_map.nodes:
+            #     for node in row:
+            #         for connected_node in node.connected_previous_row_nodes:
+            #             # get positions
+            #             node_centre_x = node.pos[0] + (node_width / 2)
+            #             node_top = node.pos[1] - line_offset
+            #             connected_node_centre = connected_node.pos[0] + (node_width / 2)
+            #             connected_node_bottom = connected_node.pos[1] + node_height + line_offset
+            #             # FIXME - are these nodes the wrong way round?
+            #
+            #             pygame.draw.line(
+            #                 surface,
+            #                 (255, 255, 255),
+            #                 (node_centre_x, node_top),
+            #                 (connected_node_centre, connected_node_bottom),
+            #             )
+            #
+            # # draw selection
+            # selected_node = overworld_map.nodes[overworld_map.current_node_row][self.selected_node]
+            # selected_node_centre_x = selected_node.pos[0] + (node_width / 2)
+            # selected_node_centre_y = selected_node.pos[1] + (node_height / 2)
+            # pygame.draw.circle(surface, (255, 0, 0), (selected_node_centre_x, selected_node_centre_y), node_width, 2)
 
             self.draw_instruction(surface)
